+++ Balance/BalanceDiaryFeature/ViewModels/DiaryFeatureViewModel.swift
@@ -1,0 +1,12 @@
+//
+//  DiaryFeatureViewModel.swift
+//  Balance
+//
+//  Created by Vishnu Ravi on 2/23/23.
+//
+
+import Foundation
+
+class DiaryFeatureViewModel: ObservableObject {
+    
+}
