+++ Balance/BalanceDiaryFeature/Views/PastDiaryEntry.swift
