+++ Balance/Home.swift
@@ -7,10 +7,10 @@
 //
 
 import BalanceContacts
-import BalanceMockDataStorageProvider
 import BalanceSchedule
 import BalanceSharedContext
 import SwiftUI
+import BalanceMockDataStorageProvider
 
 
 struct HomeView: View {
@@ -18,6 +18,7 @@ struct HomeView: View {
         case schedule
         case contact
         case mockUpload
+        case diary
     }
     
     
@@ -36,10 +37,10 @@ struct HomeView: View {
                 .tabItem {
                     Label("CONTACTS_TAB_TITLE", systemImage: "person.fill")
                 }
-            MockUploadList()
-                .tag(Tabs.mockUpload)
+            DiaryHomeView()
+                .tag(Tabs.diary)
                 .tabItem {
-                    Label("MOCK_UPLOAD_TAB_TITLE", systemImage: "server.rack")
+                    Label("Diary", systemImage: "book")
                 }
         }
     }
